hook up exclusion algo and deal with small boundaries
Take the first exclusion algorithm from `bubble_experimentation.ipynb`
and incorporate into the main codebase.

Also use a minimum bounding circle when the boundary is too small to
be able to contain any 1km circles.  In this case, instead use a
minimum bounding circle to cover the whole thing and rely on
exclusions to trim it down to a reasonable shape.